Provide knex database configuration file
knex.js requires a config object to determine what client adapter will
be used with it. Given that sqlite3 is being used, a file is specified
due to its vastly different connection structure from other options.

In the future, this should be handled differently in order to achieve
less coupling against any specific database client. The config file has
not been hardcoded at all in the application codebase in spirit of this.