provide an interface for interacting with entity
it is possible that the boundary will need to apply some business logic
at some point to filter some links, or to prepare them for sending
(perhaps validating links are actual links is a good idea). When this
occurs, the saveBookmark method will become more robust.

As it is currently written the request object and the object managed by
the business object do not have to be, but are coupled by request.body.
A mapper could feasibly be injected at the call layer of the boundary.