define a prototypical writeBoundary
because this is not finished, it is not exposed for consumption. But,
because field definitions are leveraged in both the boundary and the
entity (though not necessary here), there is an opportunity to reduce
the sources of truth. This does have the potential side effect of making
the resource the coupling point for everything. The architecture still
needs a little fleshing out.

The prototype takes a controller and a ruleset and leverages the
controller write on the result of the ruleset. This means that there's
an expectation of either assumed order of the rules or unified contract.