expose a read and write interface in the resource
this is a near mapping of the boundary, which trends toward the problems
of n-tier. The preferred use case would be to leverage a more dynamic
boundary that can be provided rules and entities as a part of its
creation.

In addition, the names of these methods leave something to be desired.
The output of this could be a collectio of {endpoint, http_method,
procesor} which could either be provided directly to express or
leveraged in a router.

As it is, these methods are just request masks on the boundary.