provide a pattern for ingress and egress transformation
the current boundary always applies the controller method and returns
what is produced by the controller. However, it is possible that an
egress mapper will need to strip off attributes or make modifications
on the returned object as well.

For the near term, the boundary will act as a wrapper on the composer.